# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: RH Cursos API
  version: 0.1.0
  summary: Contrato HTTP atual da plataforma RH Cursos
  description: >-
    Fundação OpenAPI 3.x baseada na superfície HTTP atualmente implementada no
    Route Handler `app/api/auth/session/route.ts` e nas Edge Functions em
    `supabase/functions/*/index.ts`.
  license:
    name: Proprietary
servers:
  - url: https://rhcursos.com.br
    description: Aplicação pública
  - url: https://{project-ref}.supabase.co
    description: Base do projeto Supabase para Edge Functions
    variables:
      project-ref:
        default: your-project-ref
...
paths:
...
  /api/auth/password-recovery:
    post:
      tags: [Auth Session]
      summary: Solicita um link de recuperação de senha
      description: >-
        Envia as instruções pelo Supabase Auth. A resposta é uniforme para
        evitar enumeração de contas; a rota aplica rate limit por endereço IP.
      operationId: requestPasswordRecovery
      security: []
      x-rate-limit:
        key: password-recovery:{ip}
        windowSeconds: 900
        maxRequests: 5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRecoveryRequest'
      responses:
        '200':
          description: Solicitação aceita sem revelar se a conta existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordRecoveryResponse'
        '400':
          $ref: '#/components/responses/LoginInvalid'
        '429':
          $ref: '#/components/responses/AuthRateLimited'
        '503':
          $ref: '#/components/responses/AuthUnavailable'
  /api/auth/password-update:
    post:
      tags: [Auth Session]
      summary: Atualiza a senha na sessão de recuperação
      description: >-
        Exige uma sessão Supabase SSR válida criada pelo callback de recuperação,
        valida a força da senha e encerra a sessão após a troca.
      operationId: updatePassword
      security:
        - supabaseSsrCookies: []
      x-rate-limit:
        key: password-update:{ip}
        windowSeconds: 900
        maxRequests: 5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordUpdateRequest'
      responses:
        '200':
          description: Senha atualizada e sessão local encerrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          $ref: '#/components/responses/LoginInvalid'
        '401':
          $ref: '#/components/responses/SsrSessionInvalid'
        '429':
          $ref: '#/components/responses/AuthRateLimited'
        '503':
          $ref: '#/components/responses/SsrAuthUnavailable'
  /api/auth/password-recovery/session:
    post:
      tags: [Auth Session]
      summary: Converte o token hash de recuperação em sessão SSR
      description: >-
        Recebe os tokens presentes no fragmento do link padrão do Supabase,
        grava a sessão em cookies httpOnly e nunca devolve os tokens ao cliente.
      operationId: establishPasswordRecoverySession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRecoverySessionRequest'
      responses:
        '200':
          description: Sessão de recuperação estabelecida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          $ref: '#/components/responses/LoginInvalid'
        '401':
          $ref: '#/components/responses/SsrSessionInvalid'
        '429':
          $ref: '#/components/responses/AuthRateLimited'
        '503':
          $ref: '#/components/responses/SsrAuthUnavailable'
  /api/auth/password-recovery/confirm:
    post:
      tags: [Auth Session]
      summary: Confirma o código de recuperação e cria sessão SSR
      description: Aceita o código PKCE ou token hash do e-mail de recuperação.
      operationId: confirmPasswordRecovery
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordRecoveryConfirmRequest'
      responses:
        '200':
          description: Sessão de recuperação estabelecida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '400':
          $ref: '#/components/responses/LoginInvalid'
        '401':
          $ref: '#/components/responses/SsrSessionInvalid'
        '429':
          $ref: '#/components/responses/AuthRateLimited'
        '503':
          $ref: '#/components/responses/SsrAuthUnavailable'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rh_cursos_demo_session
      description: Cookie HttpOnly emitido por `/api/auth/session`
    trustedBffServiceRoleBearer:
      type: http
      scheme: bearer
      description: >-
        Header `authorization: Bearer <SUPABASE_SERVICE_ROLE_KEY>` usado
        exclusivamente pelo BFF same-origin. Deve ser enviado em conjunto com
        `apikey` contendo a mesma service-role key; browsers não chamam
        `admin-resources` diretamente e nenhuma sessão HMAC é aceita.
    trustedBffServiceRoleApiKey:
      type: apiKey
      in: header
      name: apikey
      description: >-
        Header `apikey: <SUPABASE_SERVICE_ROLE_KEY>` exigido por
        `requireTrustedSsrAdmin` junto com `authorization` bearer da mesma
        service-role key.
    supabaseSsrCookies:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: >-
        Cookies httpOnly de sessão Supabase gerenciados por `@supabase/ssr`,
        enviados automaticamente same-origin. Nenhum token é lido do corpo.
...
  responses:
...
    LoginInvalid:
      description: Dados de login ausentes ou role inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid:
              value:
                ok: false
                error: Dados de login invalidos.
...
    AuthRateLimited:
      description: Limite de tentativas de autenticação excedido
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            limited:
              value:
                ok: false
                error: Muitas tentativas. Tente novamente mais tarde.
...
    AuthUnavailable:
      description: Ambiente sem credenciais Supabase configuradas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unavailable:
              value:
                ok: false
                error: Auth indisponivel.
    SsrSessionInvalid:
      description: Sessão SSR ausente, inválida ou expirada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid:
              value:
                ok: false
                error: Sessao invalida ou expirada.
...
    SsrAuthUnavailable:
      description: Sessão SSR do Supabase não configurada no ambiente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unavailable:
              value:
                ok: false
                error: Auth indisponivel.
...
  schemas:
...
    OkResponse:
      type: object
      additionalProperties: true
      properties:
        ok:
          type: boolean
          const: true
      required: [ok]
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: string
      required: [ok, error]
...
    PasswordRecoveryRequest:
      type: object
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
      required: [email]
    PasswordRecoveryResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        message:
          type: string
      required: [ok, message]
    PasswordUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        password:
          type: string
          minLength: 12
        confirmation:
          type: string
          minLength: 12
      required: [password, confirmation]
    PasswordRecoverySessionRequest:
      type: object
      additionalProperties: false
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [accessToken, refreshToken]
    PasswordRecoveryConfirmRequest:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
        tokenHash:
          type: string
      minProperties: 1
...